requirements.txt
backoff==2.2.1
boto3==1.35.61
botocore==1.35.61
code_bert_score==0.4.1
datasets==3.1.0
docker_py==1.10.6
google-genai>=1.0.0
litellm==1.52.8
openai==1.54.4
python-dotenv==1.0.1
Requests==2.32.3
swebench==2.1.2
tqdm==4.67.0
